Move predicates to their own file
