+++ packages/miniplex-core/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./archetypes"
 export * from "./Bucket"
+export * from "./predicates"
 export * from "./types"
 export * from "./World"
